# IGNITE-23682 Sql. Avoid unnecessary task scheduling for simple plan