https://issues.apache.org/jira/browse/IGNITE-23682

Results on MBP M3 as follow:
```
Before
Benchmark               (clusterSize)  (fsync)  Mode  Cnt  Score   Error  Units
SelectBenchmark.kvGet               1    false  avgt   20  3.025 ± 0.126  us/op
SelectBenchmark.sqlGet              1    false  avgt   20  8.652 ± 0.142  us/op

After
Benchmark               (clusterSize)  (fsync)  Mode  Cnt  Score   Error  Units
SelectBenchmark.sqlGet              1    false  avgt   20  5.872 ± 0.098  us/op



Before
Benchmark                          (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt   Score   Error  Units
InsertBenchmark.kvInsert                       1    false                24               1  avgt   20  38.814 ± 9.475  us/op
InsertBenchmark.sqlPreparedInsert              1    false                24               1  avgt   20  57.387 ± 2.757  us/op

After
Benchmark                          (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt   Score   Error  Units
InsertBenchmark.sqlPreparedInsert              1    false                24               1  avgt   20  43.395 ± 9.458  us/op
```

-----------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)